Main program code/server/website/requirements.txt
numpy
matplotlib
pillow
mysql.connector.python
scipy
django
whitenoise
python-statemachine